RESTful  不支持批量插入的提示。  #237
